Fixed bug description WRT Safari version
Bug description made it sound like Safari did not currently support IndexedDB instead Web Workers, which was fixed ~v 10.  Not sure of the exact version, but release 10 lines up with the dates on the bug ticket.  It might be better to adjust compatibility to 10.1, as IndexedDB is not mentioned in the [release notes](https://developer.apple.com/library/archive/releasenotes/General/WhatsNewInSafari/Articles/Safari_10_0.html#//apple_ref/doc/uid/TP40014305-CH11-SW1). Or someone with a mac could actually test this....